# apps/web/src/generated/openapi-schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Reframe V7 API",
    "description": "Dual-Run Retrospective Deep Narrative Forensics API with ClickHouse & Google ADK",
    "version": "7.0.0"
  },
  "paths": {
...
    "/api/v1/moments/{moment_id}/quick-wow": {
      "get": {
        "tags": [
          "Fan Experience"
        ],
        "summary": "Get Quick Wow",
        "description": "1. Quick Wow: Mobile-first visual contrast card.",
        "operationId": "get_quick_wow_api_v1_moments__moment_id__quick_wow_get",
        "parameters": [
          {
            "name": "moment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Moment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/moments/{moment_id}/rewatch-sync": {
      "get": {
        "tags": [
          "Fan Experience"
        ],
        "summary": "Get Rewatch Sync",
        "description": "2. Rewatch Sync: Dual-stream synchronized video player cuepoints and A/B loops.",
        "operationId": "get_rewatch_sync_api_v1_moments__moment_id__rewatch_sync_get",
        "parameters": [
          {
            "name": "moment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Moment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/moments/{moment_id}/trust-label": {
      "get": {
        "tags": [
          "Fan Experience"
        ],
        "summary": "Get Trust Label",
        "description": "3. Trust Label: Scientific transparency breakdown of narrative evidence grounding.\nEnforces ViewerContext spoiler protection.",
        "operationId": "get_trust_label_api_v1_moments__moment_id__trust_label_get",
        "parameters": [
          {
            "name": "moment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Moment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/moments/{moment_id}/deep-dive": {
      "get": {
        "tags": [
          "Fan Experience"
        ],
        "summary": "Get Deep Dive",
        "description": "4. Deep Dive: Forensic workbench with complete graph path and counterfactual data.",
        "operationId": "get_deep_dive_api_v1_moments__moment_id__deep_dive_get",
        "parameters": [
          {
            "name": "moment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Moment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}